+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed error in document symbols not conforming to specification - `selectionRange` will now be fully enclosed by `range`
+
 ## [1.16.0] - 2023-01-29
 
 ### Added